requirements.txt
PyMuPDF==1.26.3
openai==1.96.1
requests==2.31.0
pandas==2.2.3
python-dotenv==1.0.0
python-docx==1.2.0
